requirements.txt
pandas
jupyter
boto3
pyarrow
pyodbc
fastapi
uvicorn[standard]
pydantic
azure.storage.blob
azure.identity
sqlalchemy
python-dotenv